fix: Table widget default selected row should display value "0" instead of a watermark (#16626)
* Changed placeholderText to defaultValue

To resolve this issue I swapped out two properties of this object, placeholderText and defaultValue, to change the placeholder zero for an actual 0 character.

* Changed defaultValue from a string to an int

It seems there was an issue being caused by having the default value being a string where typing a zero still resulted in a value of -1. Changing it to an int seems to have resolved this issue.