fix: Checkbox group widget fails to display the updated values  (#19504)
## Description

In the issue report by the user the component was not checked values for
defaultSelectedValues. That was because we had componentDidUpdate which
was checking the change in the options property. If there was a change,
it was setting the selectedValues. We found out that this check is not
required. Because of this check only, the component was re-rendering,
and sometimes it was unchecking all the checked values. So we just
removed the check and added the cypress check.

Fixes #19474

## Type of change
- Bug fix 


## How Has This Been Tested?
- Cypress

### Test Plan
> https://github.com/appsmithorg/TestSmith/issues/2143

### Issues raised during DP testing
> Link issues raised during DP testing for better visiblity and tracking
(copy link from comments dropped on this PR)


## Checklist:
### Dev activity
- [ ] My code follows the style guidelines of this project
- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] New and existing unit tests pass locally with my changes
- [ ] PR is being merged under a feature flag


### QA activity:
- [x] Test plan has been approved by relevant developers
- [x] Test plan has been peer reviewed by QA
- [ ] Cypress test cases have been added and approved by either SDET or
manual QA
- [ ] Organized project review call with relevant stakeholders after
Round 1/2 of QA
- [ ] Added Test Plan Approved label after reveiwing all Cypress test