fix: flaky tests due to startup happening async (#20079)
Testcases were flaky because loading cache after application is ready is
happening in async with testcases, so sometimes testcases fails if cache
is not loaded by the time testcase executes.
Testcases fail and gives the following error message `Appsmith server is
not ready. Please try again in some time`

This fixes the issue by running startup process in sync after
application is ready.

And there is a genuine failure of testcase
`exampleApplicationsAreMarked` which is fixed.