feat: Add `instanceId` in tenant response (#20039)
In the tenant response, this PR adds the current `instanceId` to the
tenant response.

Why do we need it? This will be used to power the `Upgrade` buttons in
Admin Settings to start the U&B flow.

Why add it to the tenant API? So that for U&B flows on the cloud, we can
add something like a `tenantId` there, in the same request, so this felt
like the right place.

Potential performance impact? Unlikely. The `instanceId` is permanently
cached in the process memory, so this is a one-time DB-call, in the
lifetime of the server process.