chore: Remove usage pulse logging for Appsmith cloud (#20588)
## Description

> We are working on usage and billing feature and as a part of usage
calculation, we are recording the usage pulses at 1 hr intervals today.
As the business cases are evolving we received a request to log the
pulses for every 5mins instead of an hour. Just to provide the existing
stats for the past month on our cloud alone we have received 155545
pulses with 1hr intervals between the pulses. Also going forward if are
doing it with 5mins intervals this count is going to increase by 12
folds. Now considering we will not be monetising the cloud platform
until multi-tenancy is out, do we have usecase where we still need the
usage data? If not we would like to stop recording these usage pulses on
Appsmith cloud and bring it back after multitenancy is live. Until then
self-hosted EE instances will report the pulses as expected.

> TL;DR: Remove logging of usage pulses for Appsmith cloud not to bloat
the cloud services DB, as these pulses will not be used unless the
multi-tenancy is introduced

Fixes https://github.com/appsmithorg/cloud-services/issues/400

## Type of change

- Chore (housekeeping or task changes that don't impact user perception)

## How Has This Been Tested?

- Manual
- JUnit

## Checklist:
### Dev activity
- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] New and existing unit tests pass locally with my changes
- [x] PR is being merged under a feature flag


### QA activity:
- [ ] Test plan has been approved by relevant developers
- [ ] Test plan has been peer reviewed by QA
- [ ] Cypress test cases have been added and approved by either SDET or
manual QA
- [ ] Organized project review call with relevant stakeholders after
Round 1/2 of QA
- [ ] Added Test Plan Approved label after reveiwing all Cypress test