fix: Table breaks on adding binding to column name (#20686)
## Description

When user adds any sort of binding to the column name of a table from
property pane Columns section, the table breaks.
This was happening because of the column label getting stored as
undefined or as an object that the binding evaluates to.
Only string values are expected in column label. Hence the table was
breaking for non string values of column label.

This PR fixes this issue by replacing the column name with a default
text "customColumn" whenever the column label is not of type string

https://www.loom.com/share/1c5feab1aafa43abaac757d853a02770

Fixes #20663 

## Type of change

- Bug fix (non-breaking change which fixes an issue)


## How Has This Been Tested?

- Cypress

### Test Plan
> https://github.com/appsmithorg/TestSmith/issues/2188

### Issues raised during DP testing
> Link issues raised during DP testing for better visiblity and tracking
(copy link from comments dropped on this PR)


## Checklist:
### Dev activity
- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] New and existing unit tests pass locally with my changes
- [ ] PR is being merged under a feature flag


### QA activity:
- [x] Test plan has been approved by relevant developers
- [x] Test plan has been peer reviewed by QA
- [ ] Cypress test cases have been added and approved by either SDET or
manual QA
- [ ] Organized project review call with relevant stakeholders after
Round 1/2 of QA
- [ ] Added Test Plan Approved label after reveiwing all Cypress test