chore: Remove collaboration pointers from the Editor (#19660)
## Description

This is an internal cleanup change, which removes the collaboration
pointers from the Editor.

**_Background_**: We had a feature for collaboration which we had
sunset. As a part of this feature, we could see the mouse pointers of
the other users who were editing the same page in realtime. This feature
is today visible only to internal Appsmith users. This part of the
feature, also doesn't serve a purpose for our users.

**_Why are we removing this?:_** This code has no impact on our users.
This makes maintaining this code, as well as the additional
computational load on the components unnecessary.

**_How do we put this back in?:_** The idea of this PR is to see the
changes in one place and hence make it easy to revert when needed.

Fixes #19659 19659

Media
Here is what it looked like:
![Screenshot 2023-01-10 at 2 04 42
PM](https://user-images.githubusercontent.com/103687/211501101-465d00d8-f094-429f-83a2-c96ee3b2c76c.png)




Co-authored-by: Shrikant Sharat Kandula <shrikant@appsmith.com>