fix: Fix test email not respecting SSL option (#17952)
Fixes #16930
Fixes #16273
Fixes #20884

For connecting to some SMTP servers with TLS, we need the
`mail.smtp.ssl.enable` property to be set to `true`, in addition to the
`mail.smtp.starttls.enable` property. We add that in this PR.

Also, the username/password fields are now _always_ visible. Previously,
if you wanted to configure credentials, with TLS disabled, you have to
enable TLS, set credentials, and then disable TLS again.

Also, the send-test-email API was ignoring the TLS setting in Admin
Settings. This PR fixes that as well, but including that piece of
information in the payload sent to the server, and the API to make use
of it.