fix: widget borders does not reflect widget height (#23694)
## Description

#### Reported issue
If a widget - not a container like widget - has a minimum height of 4
rows, and the auto height is enabled for the widget, the widget borders
signify a much smaller height. Example - checkbox widget.

#### Results of RCA
This was happening because, we set `height: auto` for all widgets, when
it should have been done only for container like widgets. The
`height:auto` in the resizable library, allowed the library to show the
widget borders completely hugging the contents of the widget, when it
should have been restricted to a minimum of 4 rows (40px). This is
because, in the platform, the minimum height any widget can have is 4
rows (40px) in Fixed mode (not auto layout mode)


#### Solution:
We exclude widgets which do not have `isCanvas` property set to `true`
from setting the `autoHeight` flag. This allows the library to set the
provided height instead of hugging contents.

#### PR fixes following issue(s)
Fixes #23690 

#### Type of change
- Bug fix (non-breaking change which fixes an issue)
## Testing
#### How Has This Been Tested?
- [x] Manual
- [x] Cypress