chore: Remove unused, misleading response to PUT /env APIs (#26431)
This response object to `PUT /env` API appears to indicate whether
restart is required or not, because it includes the field
`isRestartRequired`, but this is actually not used at all on the client.
We might've used it in the past, but we don't anymore. This is a
misleading response DTO, and it's already wasted my time thrice! 😭