fix: JUnit testcases failures for EE codebase (#26824)
## Description
As we have started using the @FeatureFlagged for selectively calling the
methods from different classes based on the feature flag support, this
needs a valid user object to detect if the feature is supported for the
current user. This PR adds the api_user context for running the
testcases and also adds the mocking for FeatureFlagService which will be
used only in EE codebase, but added here to avoid any future merge
conflicts.