fix: Stop the "could not map" flood of logs (#30290)
With the recent changes to the boolean `deleted` field, we have a flood
of logs like this:

```
Could not map 'Domain.deleted'; Maybe a fragment in 'Boolean' is considered a simple type; Mapper continues with deleted
```

Where it's repeated for several of the domain classes, in place of
`Domain`.

The logs don't seem to have any functional impact anywhere, but are a
very loud noise. This PR fixes this noise.