chore: Remove non-static `or` API in Bridge (#31678)
Removing the `BridgeQuery.or` API since

1. it reads confusing. Are the passed-in items applied with an `or`, or
the current queries applied with the parameters with an `or`. Unlike
`and` method, this difference with `or` can change the meaning of the
query drastically.
2. it doesn't translate very well into Postgres world. Multiple hoops to
manage the same API, just not worth it.

The static `Bridge.or` looks, reads, and works much better. That's the
suggested alternative for this.