chore: Simpler sequence implementation (#32264)
Don't use MongoDB API to decide the key name to use for the sequence
name. This API is currently used only for datasources, so this
implementation will be compatible with both MongoDB and Postgres.

The actual sequence number generation, in the `getNext` method, will
basically need to be rewritten in the Postgres branch.