chore: Remove API to get all objects (#33006)
The `.get(params)` method in `BaseService` has a dangerous default
implementation. It loads _all_ objects from the DB and returns them.
Although it's not used by any services in that form, it's nevertheless a
liability, and not an asset.

For example, the User service explicitly overrides it and always
responds with an error.

This PR removes this method, and we add the signature to just those
services that leverage it.

EE at https://github.com/appsmithorg/appsmith-ee/pull/3698. Tests and
Cypress pass.

/ok-to-test tags="@tag.Sanity"