fix: Remove redundant action collection update call on page load in edit mode (#33139)
## Description
If there are any functions in a JS Object that accept an argument, an
update action collection for the JavaScript object is triggered when the
application first loads. These update calls alter the lastUpdatedAt time
of a page merely by opening the application in edit mode, without making
any changes.

In getDifferenceInJSCollection, we check if the arguments to the action
(function in a JS object) are the same as those in the Redux store. This
check is not required as we are already checking the equality of the
body and arguments are also part of the body.

This check was always returning false because one of the object has an
extra value key
```
x = [{ name: "argument1" }]
y = [{ name: "argument1" , value: undefined }]
```

Fixes #`Issue Number`  
_or_  
Fixes `Issue URL`
> [!WARNING]  
> _If no issue exists, please create an issue first, and check with the
maintainers if the issue is valid._

## Automation

/ok-to-test tags="@tag.All"

### :mag: Cypress test results
<!-- This is an auto-generated comment: Cypress test results  -->
> [!TIP]
> 🟢 🟢 🟢 All cypress tests have passed! 🎉 🎉 🎉
> Workflow run:
<https://github.com/appsmithorg/appsmith/actions/runs/8967578757>
> Commit: 5b1cdb75986374be3469b22382c542bf123a8ead
> Cypress dashboard url: <a
href="https://internal.appsmith.com/app/cypress-dashboard/rundetails-65890b3c81d7400d08fa9ee5?branch=master&workflowId=8967578757&attempt=2"
target="_blank">Click here!</a>

<!-- end of auto-generated comment: Cypress test results  -->











## Communication
Should the DevRel and Marketing teams inform users about this change?
- [ ] Yes
- [x] No