chore: Remove incorrect and unused `update` signature (#33529)
The `update` method with `String, T, String` signature, is not used in
it's form anywhere. This is most likely owing to the implementation
being off. The criteria used to find the object to be updated, is not
the same as the criteria used to find the updated object to return to
the caller.

The query to update is `{value of key field} = id`.

The query to find is `{value of "id" field} = id`.

This discrepency is likely why this function is not used anywhere. The
`ApplicationServiceCE` alone has an overridden implementation for this
signature, but that's defined with different variable names, and treats
the third argument as the Git branch name, not `key`. That function is
indeed used, and is unaffected by this change.

All of above is verified on EE as well. No conflicts, no build failure.

/ok-to-test tags="@tag.Sanity"