chore: add new custom linter for module inputs (#33940)
## Description
This PR adds a new custom lint function for module inputs. The primary
logic for linting would be present in the EE counterpart. This PR makes
sure the right parameter is passed to the right function and the linter
function is appropriately split for EE

It also a generic `CustomLintErrorCode` called `INVALID_INPUTS`; this is
made generic as there is a possibility of introducing inputs to normal
query and hence this can be reused. If that doesn't happen, it still
should be fine as inputs would only be a concept reserved to module.

PR for https://github.com/appsmithorg/appsmith-ee/pull/4339

## Automation

/ok-to-test tags="@tag.All"

### :mag: Cypress test results
<!-- This is an auto-generated comment: Cypress test results  -->
> [!TIP]
> 🟢 🟢 🟢 All cypress tests have passed! 🎉 🎉 🎉
> Workflow run:
<https://github.com/appsmithorg/appsmith/actions/runs/9382849520>
> Commit: 18260e82e9816daf4e6be49a83c0fb08dfc491a4
> Cypress dashboard url: <a
href="https://internal.appsmith.com/app/cypress-dashboard/rundetails-65890b3c81d7400d08fa9ee5?branch=master&workflowId=9382849520&attempt=2"
target="_blank">Click here!</a>

<!-- end of auto-generated comment: Cypress test results  -->









## Communication
Should the DevRel and Marketing teams inform users about this change?
- [ ] Yes
- [ ] No


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Improved error handling for invalid module inputs with detailed error
messages.

- **Enhancements**
- Enhanced linting functionality by adding checks for invalid module
inputs.

- **Bug Fixes**
- Fixed inconsistencies in error message generation for invalid inputs.

- **Refactor**
- Renamed and updated parameter types for improved code clarity and
maintainability.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->