ci: Add a pre-push hook to check for only ce changes (#34992)
## Description
This PR checks for file changes when pushing in CE repo and passes if
the changes are present only in CE folder. To override the check,
developers can use --no-verify check when pushing the changes. This PR
is an extension of https://github.com/appsmithorg/appsmith/pull/32791


Fixes #`Issue Number`  
_or_  
Fixes `Issue URL`
> [!WARNING]  
> _If no issue exists, please create an issue first, and check with the
maintainers if the issue is valid._

## Automation

/ok-to-test tags=""

### :mag: Cypress test results
<!-- This is an auto-generated comment: Cypress test results  -->
> [!CAUTION]  
> If you modify the content in this section, you are likely to disrupt
the CI result for your PR.

<!-- end of auto-generated comment: Cypress test results  -->


## Communication
Should the DevRel and Marketing teams inform users about this change?
- [ ] Yes
- [x] No


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Chores**
- Implemented a pre-push Git hook to check for specific file changes and
provide feedback before pushing to a remote repository.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->