chore: add rollup config for wds (#39397)
With this PR, we want to publish WDS as a npm package. To do that, we
are now adding a rollup setup to compile and build the package.

One thing to note:
Rollup needed `7.24.0` version of browserlist and we had to update
browserlist version in resolutions in root package.json. Now since we
updated browserlist, certain code that was written with old format (
code was coming from blueprint's hotkeys component ) started failing and
we had to refactor. It included the code around hotkeys component. It
was deprecatdd by blueprint and we were still using. We have now
refactored those part of code to use the `useHotKeys` hook.

/ok-to-test tags="@tag.All"

<!-- This is an auto-generated comment: Cypress test results  -->
> [!TIP]
> 🟢 🟢 🟢 All cypress tests have passed! 🎉 🎉 🎉
> Workflow run:
<https://github.com/appsmithorg/appsmith/actions/runs/13538917574>
> Commit: 9b0b51791e4a95574c9729245ba09994ab371b71
> <a
href="https://internal.appsmith.com/app/cypress-dashboard/rundetails-65890b3c81d7400d08fa9ee5?branch=master&workflowId=13538917574&attempt=2"
target="_blank">Cypress dashboard</a>.
> Tags: `@tag.All`
> Spec:
> <hr>Wed, 26 Feb 2025 09:12:45 UTC
<!-- end of auto-generated comment: Cypress test results  -->